+++ src/monsterviews.h
@@ -1,0 +1,60 @@
+/*
+ *	Monster Chase: a testing playground for behaviors trees
+ *
+ *	Copyright 2019 Andrea Ravalli <anravalli @ gmail.com>
+ *
+ *	This file is part of Monster Chase.
+ *
+ *	Monster Chase is free software: you can redistribute it and/or modify
+ *	it under the terms of the GNU General Public License as published by
+ *	the Free Software Foundation, either version 3 of the License, or
+ *	(at your option) any later version.
+
+ *	Monster Chase is distributed in the hope that it will be useful,
+ *	but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *	GNU General Public License for more details.
+
+ *	You should have received a copy of the GNU General Public License
+ *	along with Monster Chase.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef MONSTERVIEWS_H
+#define MONSTERVIEWS_H
+
+#include <QtWidgets>
+#include <QGraphicsItem>
+
+#include "monster.h"
+
+namespace Monster {
+
+class MonsterShape : public QGraphicsItem
+{
+public:
+    MonsterShape(MonsterModel* m);
+    QRectF boundingRect() const Q_DECL_OVERRIDE;
+    void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget) Q_DECL_OVERRIDE;
+
+private:
+    QColor color[3];
+    MonsterModel* model;
+};
+
+class MonsterSight : public QGraphicsItem
+{
+public:
+    MonsterSight(MonsterModel* m)
+        :model(m){}
+
+    QRectF boundingRect() const Q_DECL_OVERRIDE;
+
+    void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget) Q_DECL_OVERRIDE;
+
+private:
+    MonsterModel* model;
+};
+
+}
+
+#endif // MONSTERVIEWS_H
